move error handling to a new file
